Refactor Time + Zone avec Upload de fichier
