+++ app/models/Application.scala
@@ -85,7 +85,16 @@ case class Application(id: UUID,
      (user.expert && invitedUsers.keys.toList.contains(user.id) && !closed)||
      creatorUserId==user.id
 
-  def fileCanBeShowed(user: User) = user.instructor && invitedUsers.keys.toList.contains(user.id)
+  def fileCanBeShowed(user: User, answer: UUID) =
+    answers.find(_.id == answer) match {
+      case None => false
+      case Some(answer) if answer.filesAvailabilityLeftInDays != None => false // You can't download expired file
+      case Some(answer) if answer.creatorUserID == user.id => false   // You can't download your own file
+      case _ =>
+        (user.instructor && invitedUsers.keys.toList.contains(user.id)) ||
+          (user.helper && user.id == creatorUserId)
+    }
+
 
   def canBeAnsweredBy(user: User) =
     (user.instructor && invitedUsers.keys.toList.contains(user.id)) ||
