+++ app/models/LoginToken.scala
@@ -2,6 +2,7 @@ package models
 
 import java.util.UUID
 
+import extentions.Time
 import org.joda.time.DateTime
 
 import scala.util.Random
