+++ app/views/showApplication.scala.html
@@ -1,4 +1,4 @@
-@(user: User, area: Area)(users: List[User], application: Application, answerToAgentsForm: Form[_])(implicit webJarsUtil: org.webjars.play.WebJarsUtil, flash: Flash, messagesProvider: MessagesProvider)
+@(user: User, area: Area)(users: List[User], application: Application, answerToAgentsForm: Form[_], attachmentsActivated: Boolean)(implicit webJarsUtil: org.webjars.play.WebJarsUtil, flash: Flash, messagesProvider: MessagesProvider)
 @import java.util.Locale
 
 @main(user, area)(s"Demande : ${application.subject} ")  {
@@ -263,7 +263,7 @@ dialog {
                             font-weight: 600;
                             line-height: 16px;">
                             <i class="icon material-icons icon--light">attach_file</i>
-                            @answer.creatorUserName a ajouté le fichier @if(application.fileCanBeShowed(user) && answer.filesAvailabilityLeftInDays != None) { <a href="@routes.ApplicationController.file(application.id, answer.id, filename)" target="_blank">@filename</a> } else { @filename }
+                            @answer.creatorUserName a ajouté le fichier @if(application.fileCanBeShowed(user, answer.id)) { <a href="@routes.ApplicationController.file(application.id, answer.id, filename)" target="_blank">@filename</a> } else { @filename }
                             @answer.filesAvailabilityLeftInDays match {
                                 case None => { ( Fichier expiré et supprimé ) }
                                 case Some(expirationInDays) => { ( Suppression du fichier dans @expirationInDays jours ) }
@@ -382,7 +382,7 @@ dialog {
                                         <textarea class="mdl-textfield__input" type="text" rows= "5" id="helper-answer-message" style="width: 100%;" name="message" onkeyup="document.getElementById('review-validation').disabled=(this.value === '')"></textarea>
                                         <label class="mdl-textfield__label" for="helper-answer-message"><i class="material-icons" style="vertical-align: middle;">message</i> Votre réponse</label>
                                     </div>
-                                    @if(user.helper) {
+                                    @if(attachmentsActivated) {
                                         <div class="mdl-cell mdl-cell--12-col">
                                              <b>Ajouter un ficher ( 5Mo maximum ) : <input type="file" name="file"></b>
                                         </div>
